requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-groq==0.0.1
langchain-pinecone==0.0.3
pinecone-client==3.1.0
pypdf==4.1.0
sentence-transformers==2.5.1
python-dotenv==1.0.1
